% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatetable.R
\name{update_download_state}
\alias{update_download_state}
\title{Update data info of SRA summary tables}
\usage{
update_download_state(
  cells_table = "version3/WGS_metadata.cells_v3.txt",
  accessions_table = "version3/WGS_metadata.accessions_v3.txt",
  check_vdb,
  datasetid,
  data_location_dir
)
}
\arguments{
\item{cells_table}{cells_table path}

\item{accessions_table}{accessions_table path}

\item{check_vdb}{result from vdb-validate}

\item{datasetid}{a datasetid}

\item{data_location_dir}{location of data,end by '/'}
}
\value{
a updated table
}
\description{
check_vdb can be generates by `vdb-validate path /storage/SRP010990 2> check_vdb_mini.log`
}
\examples{
\dontrun{
update_download_state(
cells_table = "version3/WGS_metadata.cells_v3.txt",
accessions_table = "version3/WGS_metadata.accessions_v3.txt",
check_vdb = "H:/xiayonghe/check_vdb_mini.log",
datasetid = "SRA056303",
data_location_dir = "/storage/douyanmeiLab/dingyihang/SRA056303/"
)
}
}
